fixes use of Persistance
