+++ frontend/javascripts/dashboard/dataset/dataset_cache_provider.js
@@ -1,14 +1,11 @@
 // @flow
-import React, { createContext, useState, useEffect, type Node } from "react";
-import { useHistory } from "react-router-dom";
-import { PropTypes } from "@scalableminds/prop-types";
+import React, { createContext, useState, type Node } from "react";
 
+import type { DatasetFilteringMode } from "dashboard/dataset_view";
 import type { APIMaybeUnimportedDataset } from "admin/api_flow_types";
 import { getDatastores, triggerDatasetCheck, getDatasets } from "admin/admin_rest_api";
-import { type DatasetFilteringMode } from "dashboard/dataset_view";
 import { handleGenericError } from "libs/error_handling";
 import UserLocalStorage from "libs/user_local_storage";
-import Persistence from "libs/persistence";
 import * as Utils from "libs/utils";
 
 type Context = {
@@ -17,25 +14,6 @@ type Context = {
   checkDatasets: () => Promise<void>,
   fetchDatasets: (datasetFilteringMode?: DatasetFilteringMode) => Promise<void>,
 };
-
-type PersistenceState = {
-  datasets: Array<APIMaybeUnimportedDataset>,
-  isLoading: boolean,
-  searchQuery: string,
-  datasetFilteringMode: DatasetFilteringMode,
-};
-
-const persistence: Persistence<PersistenceState> = new Persistence(
-  {
-    searchQuery: PropTypes.string,
-    datasetFilteringMode: PropTypes.oneOf([
-      "showAllDatasets",
-      "onlyShowReported",
-      "onlyShowUnreported",
-    ]),
-  },
-  "datasetList",
-);
 
 const wkDatasetsCacheKey = "wk.datasets";
 export const datasetCache = {
@@ -60,7 +38,6 @@ export const DatasetCacheContext = createContext<Context>({
 export default function DatasetCacheProvider({ children }: { children: Node }) {
   const [datasets, setDatasets] = useState(datasetCache.get());
   const [isLoading, setIsLoading] = useState(false);
-  const history = useHistory();
   async function fetchDatasets(
     datasetFilteringMode?: DatasetFilteringMode = "onlyShowReported",
   ): Promise<void> {
@@ -97,17 +74,6 @@ export default function DatasetCacheProvider({ children }: { children: Node }) {
       setIsLoading(false);
     }
   }
-  useEffect(() => {
-    try {
-      persistence.load(history);
-    } catch (error) {
-      console.error(error);
-      // An unknown error was thrown. To avoid any problems with the caching of datasets,
-      // we simply clear the cache for the datasets and re-fetch.
-      setDatasets([]);
-      datasetCache.clear();
-    }
-  }, []);
   return (
     <DatasetCacheContext.Provider value={{ datasets, isLoading, checkDatasets, fetchDatasets }}>
       {children}
